+++ src/extensions/mvs/tree/mvs/param-types.ts
@@ -37,6 +37,9 @@ export const ComponentExpressionT = partial({
     end_label_seq_id: int,
     beg_auth_seq_id: int,
     end_auth_seq_id: int,
+    label_comp_id: str,
+    auth_comp_id: str,
+    // residue_index: int, // 0-based residue index in the source file // TODO this is defined in Python builder but not supported by Molstar yet
     label_atom_id: str,
     auth_atom_id: str,
     type_symbol: str,